store/wizard: move timezone list to system constants
